cashback reward

+++ src/Shopper/Customer.java
@@ -9,7 +9,7 @@ public class Customer {
     private double balance;
     private String pinCode; // New field for the 4-digit PIN code
     private double transaction;
-    private int reward_point;
+    private double reward_point;
 
 
     public Customer() {
@@ -74,12 +74,12 @@ public class Customer {
         this.transaction = transaction;
     }
 
-    //every 50pesos, customer earn 1pt
-    public int getRewardPoint() {
+    //every 100 pesos, customer earn 1pt
+    public double getRewardPoint() {
         return reward_point;
     }
 
-    public void setRewardPoint(int reward_point) {
+    public void setRewardPoint(double reward_point) {
         this.reward_point = reward_point;
     }
